added json file for data and fixed eslint in tpo page
